LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-proximity-expected.txt


FAIL Alternating light/dark assert_equals: expected "rgb(100, 100, 100)" but got "rgb(200, 200, 200)"
FAIL Proximity wins over order of appearance assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS Specificity wins over proximity
